requirements.txt
asgiref==3.2.3
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-static==0.0.6
Django==3.0.2
django-ckeditor==5.8.0
django-import-export==2.0
django-js-asset==1.2.2
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
pytz==2019.3
PyYAML==5.2
sqlparse==0.3.0
static3==0.7.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
